Rename namespace to UserGroup
