+++ src/Umbraco.Cms.Api.Management/Controllers/UserGroup/UpdateUserGroupController.cs
@@ -7,7 +7,7 @@ using Umbraco.Cms.Core.Models.Membership;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Services.OperationStatus;
 
-namespace Umbraco.Cms.Api.Management.Controllers.UserGroups;
+namespace Umbraco.Cms.Api.Management.Controllers.UserGroup;
 
 public class UpdateUserGroupController : UserGroupControllerBase
 {
